Add files via upload
The RGB color game has the user guess 1 out of 3 or 6 colors depending on the difficulty of the game which the users sets themselves. The RGB value that the user has to guess is displayed at the top. This was a full project which involved javascript logic and css styling as well as an HTML structure of the page. The javascript logic relies on click listeners quite a lot, since the user will be able to interact with the application through click events. Each button corresponds to a click event that can change the state of the game.